Edits to the Board
I have added a position element to each piece, which we can keep track
of both on the board as well as the piece. Not sure how necessary it is,
but it would be useful to keep both. I will send you a pull request
shortly. For you to merge this code in.