Board as a Table
Successful use of object oriented PHP to load the board as an HTML
table. Playing around with using GET to store different boards. Next
step is to get the MySQL set up and store boards in there. I will try
implementing an HTML form that users can move pieces with (for testing).