MySQL Storge of Boards!
Boards can be stored in, loaded from, and updated to a MySQL database.
Each board can be accessed with the GET parameter 'id' (e.g.
/board.php?id=5 is a board). There is a form on each page, for
demonstrative/testing purposes, where a user can enter from and to
spaces to update the database. This information will eventually be
coming from the Arduino, not direct user input.